Refactor project structure and remove deprecated packages
- Updated the go.work file to reflect new package paths, moving from ./services/auth and ./services/user to ./packages/auth, ./packages/db, ./packages/shared, and ./packages/user.
- Deleted obsolete go.mod and go.sum files from the pkg and services directories, along with related configuration and utility files.
- Removed the user service and its associated files, including configuration and monitoring implementations, to streamline the project.
