+++ go.work
@@ -2,7 +2,8 @@ go 1.25.0
 
 use (
 	./gen
-	./pkg
-	./services/auth
-	./services/user
+	./packages/auth
+	./packages/db
+	./packages/shared
+	./packages/user
 )
