+++ packages/auth/go.mod
@@ -1,0 +1,3 @@
+module github.com/yaninyzwitty/chat/packages/auth
+
+go 1.25.0
