+++ packages/db/go.mod
@@ -1,0 +1,20 @@
+module github.com/yaninyzwitty/chat/packages/db
+
+go 1.25.0
+
+require (
+	github.com/apapsch/go-jsonmerge/v2 v2.0.0 // indirect
+	github.com/datastax/astra-client-go/v2 v2.2.54 // indirect
+	github.com/datastax/cql-proxy v0.1.6 // indirect
+	github.com/datastax/go-cassandra-native-protocol v0.0.0-20220706104457-5e8aad05cf90 // indirect
+	github.com/datastax/gocql-astra v0.0.0-20250516142328-482592316433 // indirect
+	github.com/deepmap/oapi-codegen v1.12.4 // indirect
+	github.com/gocql/gocql v1.7.0 // indirect
+	github.com/golang/snappy v0.0.4 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/hailocab/go-hostpool v0.0.0-20160125115350-e80d13ce29ed // indirect
+	go.uber.org/atomic v1.8.0 // indirect
+	go.uber.org/multierr v1.7.0 // indirect
+	go.uber.org/zap v1.17.0 // indirect
+	gopkg.in/inf.v0 v0.9.1 // indirect
+)
