+++ packages/user/go.mod
@@ -1,0 +1,8 @@
+module github.com/yaninyzwitty/chat/packages/user
+
+go 1.25.0
+
+require (
+	github.com/joho/godotenv v1.5.1 // indirect
+	golang.org/x/sync v0.17.0 // indirect
+)
