+++ packages/user/main.go
@@ -14,9 +14,9 @@ import (
 	"github.com/joho/godotenv"
 	"github.com/prometheus/client_golang/prometheus"
 	userv1 "github.com/yaninyzwitty/chat/gen/user/v1"
-	"github.com/yaninyzwitty/chat/pkg/config"
-	"github.com/yaninyzwitty/chat/pkg/monitoring"
-	"github.com/yaninyzwitty/chat/user/controller"
+	"github.com/yaninyzwitty/chat/packages/shared/config"
+	"github.com/yaninyzwitty/chat/packages/shared/monitoring"
+	"github.com/yaninyzwitty/chat/packages/user/controller"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
