--- pkg/go.mod
@@ -1,34 +1,0 @@
-module github.com/yaninyzwitty/chat/pkg
-
-go 1.25.0
-
-require (
-	github.com/datastax/gocql-astra v0.0.0-20250516142328-482592316433
-	github.com/gocql/gocql v1.7.0
-	github.com/prometheus/client_golang v1.23.2
-	gopkg.in/yaml.v3 v3.0.1
-)
-
-require (
-	github.com/apapsch/go-jsonmerge/v2 v2.0.0 // indirect
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/datastax/astra-client-go/v2 v2.2.54 // indirect
-	github.com/datastax/cql-proxy v0.1.6 // indirect
-	github.com/datastax/go-cassandra-native-protocol v0.0.0-20220706104457-5e8aad05cf90 // indirect
-	github.com/deepmap/oapi-codegen v1.12.4 // indirect
-	github.com/golang/snappy v0.0.4 // indirect
-	github.com/google/uuid v1.3.0 // indirect
-	github.com/hailocab/go-hostpool v0.0.0-20160125115350-e80d13ce29ed // indirect
-	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
-	github.com/prometheus/client_model v0.6.2 // indirect
-	github.com/prometheus/common v0.66.1 // indirect
-	github.com/prometheus/procfs v0.16.1 // indirect
-	go.uber.org/atomic v1.8.0 // indirect
-	go.uber.org/multierr v1.7.0 // indirect
-	go.uber.org/zap v1.17.0 // indirect
-	go.yaml.in/yaml/v2 v2.4.2 // indirect
-	golang.org/x/sys v0.35.0 // indirect
-	google.golang.org/protobuf v1.36.9 // indirect
-	gopkg.in/inf.v0 v0.9.1 // indirect
-)
