--- pkg/util/doc.go
@@ -1,3 +1,0 @@
-// Package util contains any and all utilities
-// used by the application
-package util
