--- services/auth/go.mod
@@ -1,3 +1,0 @@
-module github.com/yaninyzwitty/chat/services/auth
-
-go 1.25.0
